requirements-dev.txt
coverage==7.13.5
eth-account==0.13.7
eth-tester==0.13.0b1
py-evm==0.12.1b1
py-solc-x==2.0.3
reportlab==4.4.4
slither-analyzer==0.11.5
web3==7.13.0
xlsxwriter==3.2.9
